move and rename

+++ tasmota/Modules/grepmodule-firmware.py
@@ -29,20 +29,23 @@ if env["PIOPLATFORM"] != "espressif32" :
                     fwp.write(msync)
 
                 if msync[0] == 0x4a and msync[1] == 0xfc and msync[2] == 0xaa and msync[3] == 0x55:
-                    start = 1
-                    #print("found start sync")
-                    dummy = fp.read(12)
-                    xname = fp.read(16)
-                    fname = xname.decode('ascii') 
-                    name = fname.replace('\x00','')
-                    mod_file = dir_path + "/" + name + ".bin"
-                    if os.path.isfile(mod_file):
-                        os.remove(mod_file)
-                    fwp = open(mod_file, "wb")
-                    fwp.write(msync)
-                    fwp.write(dummy)
-                    fwp.write(xname)
-                    size += 24
+                    arch = int.from_bytes(fp.read(4), "little")
+                    type = int.from_bytes(fp.read(4), "little")
+                    if arch == 0 and type <= 4:
+                        #print("found start sync")
+                        start = 1
+                        dummy = fp.read(4)
+                        xname = fp.read(16)
+                        fname = xname.decode('ascii') 
+                        name = fname.replace('\x00','')
+                        mod_file = dir_path + "/" + name + ".bin"
+                        if os.path.isfile(mod_file):
+                            os.remove(mod_file)
+                        fwp = open(mod_file, "wb")
+                        fwp.write(msync)
+                        fwp.write(dummy)
+                        fwp.write(xname)
+                        size += 24
 
                 if msync[0] == 0x55 and msync[1] == 0xaa and msync[2] == 0xfc and msync[3] == 0x4a:
                     start = 2
